requirements.txt
Django==3.0.5
PyMySQL==0.9.3
psycopg2>=2.7,<3.0
djangorestframework
django-rest-auth
django-bootstrap4
django-silk
